Ensure the Language Server is started with the latest config
